+++ tempest_fastapi_sdk/__init__.py
@@ -421,7 +421,7 @@ from tempest_fastapi_sdk.websockets import (
     make_websocket_router,
 )
 
-__version__: str = "0.151.1"
+__version__: str = "0.152.0"
 
 __all__: list[str] = [
     "BASE_COLUMN_ORDER",